xray_project/requirements.txt
Django>=4.0
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
django-filter>=22.1
Pillow>=9.0.0
faker>=18.0.0
elasticsearch>=9.1.0
django-elasticsearch-dsl>=9.0